LI: Centering dome test case
Added logic to put the center of the dome in the center of a cell.  This
means the dome may not be exactly centered in the entire domain, but
the center of the dome coincides with the center of a cell (specifically
the cell whose cell center is closest to the center of the domain).
This provies x- and y- axis symmetry for the dome, whereas before the
dome was only guaranteed to have symmetry about some arbitrary line.  As
part of this I also shift the origin of the x and y coordinates of the
grid to be the center of the domain.