LI: setup ic script for all EISMINT2 tests
The script for setting up test A in the previous commit has been
generalized to work for any of tests A-D and F.  These are all of the
tests described in the Payne et al. (2000) paper that do not include
basal slip.

Some of the b.c. fields are commented out because those fields do not
exist in the model yet.  Also for that reason I have not run any of the
tests.  However, I have tested the setup of the input files and it seems
to work as intended for the range of functionality that is currently
testable.