LI: circ. shelf: add dirichlet b.c. setup, fix x/y shift
This commit updates the Land Ice circular shelf test case so that it
enforces no slip basal boundary conditions in the grounded region of the domain
using the recently added Dirichlet velocity b.c. capability.  Previously this has
been done by assigning a large value of beta.

Supplying the "-b","--beta" command line option allows the old method -
the new Dirichlet b.c. method is the default.

There is also another new command line option, "-7","--7cells".  Use
this flag to create 7 grounded cells (the center cell and its 6 neigbors).  The
default is to only ground the center cell.

I've also added logic so that shifting the x/y coordinates of the domain
to put the origin at the center of the domain only occurs if it appears this
has not already been done.  Otherwise, the shift was occurring repeatedly which
was messing up the domain.  The check is fairly simple, but works for meshes
generated with periodic_hex.

Finally, I've updated the streams.landice file.