Make MISMIP mesh symmetric about x-axis
The MISMIP solution is much cleaner if the mesh is symmetric in the
y-direction.  I've made a copy of the
mark_periodic_boundaries_for_culling.py script in this directory and
modified to take an extra row off the mesh.

Note this requires careful choice of the ny option in periodic_hex to
make sure there will be an odd number of cells left after culling 3 cells.
(since periodic_hex currently requires an even number of cells in the y-dir)

Note that a simpler and more flexible solution would be to modify
periodic_hex to allow creation of non-periodic meshes, and if the mesh is
non-periodic remove the restriction of an even number of cells in the
y-dir.