Update confined shelf parameters
I discovered that we were using the wrong density for ice and ocean in
this test case, and that the bed under the grounded margins was a bit
too shallow, resulting in non-negligible slopes into the shelf.  This
commit corrects those issues so that are implementation of the test case
matches the test case description.